refactor: padroniza nomes de métodos e serviços relacionados a Mapa e ajusta transações em MapaManutencaoService
- Renomeia métodos em MapaFacade, controllers e testes para padronizar nomes (ex: obterPorCodigo -> mapaPorCodigo, listar -> todosMapas)
- Ajusta chamadas e mocks em testes para refletir os novos nomes
- Refatora MapaManutencaoService para uso consistente de @Transactional e separação de métodos de leitura e escrita
- Remove métodos duplicados e centraliza lógica de manipulação de mapas, atividades, competências e conhecimentos
- Melhora clareza e coesão dos serviços de visualização e contexto de subprocesso
- Ajusta EmailServiceTest e ProcessoNotificacaoServiceTest para refletir mudanças em dependências e mocks

+++ backend/src/test/java/sgc/processo/service/ProcessoManutencaoServicePbtTest.java
@@ -18,7 +18,6 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @Tag("PBT")
@@ -38,7 +37,7 @@ class ProcessoManutencaoServicePbtTest {
 
         when(unidadeService.unidadePorCodigo(any())).thenAnswer(inv -> {
              Unidade u = new Unidade();
-             u.setCodigo((Long) inv.getArgument(0));
+             u.setCodigo(inv.getArgument(0));
              u.setSituacao(SituacaoUnidade.ATIVA);
              u.setTipo(TipoUnidade.OPERACIONAL);
              return u;
@@ -65,7 +64,7 @@ class ProcessoManutencaoServicePbtTest {
 
         when(unidadeService.unidadePorCodigo(any())).thenAnswer(inv -> {
              Unidade u = new Unidade();
-             u.setCodigo((Long) inv.getArgument(0));
+             u.setCodigo(inv.getArgument(0));
              u.setSituacao(SituacaoUnidade.ATIVA);
              u.setTipo(TipoUnidade.OPERACIONAL);
              return u;
